dont do integer division

+++ Assets/Source/Components/Player/PlayerBehavior.cs
@@ -176,7 +176,7 @@ namespace Assets.Source.Components.Player
                 actorBehavior.Health -= baseDamage;
 
                 // Warn player if health is less than 10%
-                if (actorBehavior.Health > 0 && (actorBehavior.Health / actorBehavior.MaxHealth) < 0.1f)
+                if (actorBehavior.Health > 0 && ((float)actorBehavior.Health / actorBehavior.MaxHealth) < 0.1f)
                 {
                     Warning("Low Health");
                 }
